Backend/requirements.txt
# requirements.txt
fastapi
uvicorn
pydantic
redis
pymongo
requests
flask
pytubefix
groq
python-dotenv
ffmpeg-python
